Cleaned up project, commented out not needed functions and endpoints, foldered not used and not implemented classes
