+++ back-end/Logic/Interface/ISubjectLogic.cs
@@ -14,8 +14,10 @@ namespace Logic.Interface
         bool CreateSubject(Subject subject);
         bool EditSubject(string id, Subject newSubject);
         bool DeleteSubject(string id);
-        void AddUserToSubject(MyUser user, string subjectId);
-        void DeleteUserFromSubject(MyUser user, string subjectId);
         IQueryable<Subject> GetAllSubjectsOfYear(string yearId);
+
+        // NOT NEEDED
+        //void AddUserToSubject(MyUser user, string subjectId);
+        //void DeleteUserFromSubject(MyUser user, string subjectId);
     }
 }
