+++ back-end/Logic/Interface/Unused/IAwardLogic.cs
