+++ back-end/Logic/Interface/Unused/ITransactionLogic.cs
